dana delete fix

+++ main/views.py
@@ -109,11 +109,15 @@ def update_post(request, pk):
 @login_required(login_url='login')
 def delete_post(request, pk):
     post = get_object_or_404(Post, pk=pk)
-    if request.method == 'POST':
-        post.delete()
-        messages.add_message(request, messages.SUCCESS, 'You delete your blog')
-        return redirect('homepage')
-    return render(request, 'delete-post.html')
+    if request.user == post.user:
+        if request.method == 'POST':
+            post.delete()
+            messages.add_message(request, messages.SUCCESS, 'You delete your blog')
+            return redirect('homepage')
+        return render(request, 'delete-post.html')
+    else:
+        return HttpResponse('<h1>Вы не являетесь создателем этого поста!!!<h1>')
+
 
 
 class AddCommentView(SuccessMessageMixin, CreateView):
@@ -121,10 +125,13 @@ class AddCommentView(SuccessMessageMixin, CreateView):
     form_class = CommentForm
     template_name = 'comment.html'
     success_url = reverse_lazy('homepage')
-    success_message = 'Your comment successfully added to that post!!!'
+    succeformf_validss_message = 'Your comment successfully added to that post!!!'
 
 
     def form_valid(self, form):
         form.instance.post_id = self.kwargs['pk']
-        return super().form_valid(form)
+        return super().formf_valid(form)
+
+
+
